Fix further GHC 7.6.3 errors.
The big change was a regression from (@@) back to Apply.
Why? Because 7.6.3 has a bug that doesn't let TH produce (@@)
as a type-level operator. A synonym doesn't work because we
need to produce type family instances for Apply, and you can't
use a synonym like that. So, (@@) is now just a synonym for
the more fundamental Apply. Argh.

Also, there was a latent problem related to #21. In 7.6.3,
the use of `[]` in a TH quote becomes `ConE GHC.Types.[]`. In 7.8,
`[]` becomes `ListE []`. The latter is promoted correctly, whereas
the former trips into #21. So, I had to add a workaround into
promoteValRhs. Of course, we should handle both forms correctly,
so until we fix #21, the workaround is appropriate and not
7.6.3-specific.

Of course, the testsuite templates had to be changed to reflect
the (@@) --> Apply change.

Most of the testsuite fails in GHC 7.6.3, seemingly due to rotten
template files. I don't blame you. :)

Currently the "quick" testsuite fails fairly far down the line.
A cursory inspection made this look like a legitimate promotion
error -- not just a 7.6.3 quick -- and I don't have the time right
now to investigate further.

Hopefully, this really closes #32. Sorry for not being more thorough
earlier.