Pass through Promote.hs.
The precise details of the code generation is different from before,
so the testsuite falls over. I have yet to analyze the output to see
if it is sensible. Also, singleton generation is not yet switched over,
so some weird things happen. (Note the InsertionSort test case,
involving "if".)

But, it compiles, and given the bootstrapped nature of the library,
that means that it's close to working. More soon.