Make progress on #136.
This commit is a definite improvement w.r.t. #136, but it
doesn't fully fix it. Still outstanding is the fact that
the expression `... | otherwise = succ (toEnum (pred i))` is
just too involved for singletons to figure out how to
specialize succ.

Wait. Never mind. I've figured out how to do it.