Remove the MODERN_MTL hack
Previously, GHC jumped through several hoops to support mutiple versions of `mtl`.
But GHC 8.2 only supports `mtl-2.2.1` or later, so there's no longer any need for
this hackery. This allows for substantial simplication of the test suite code, and
it allows us to remove a dependency on `Cabal` in the test suite.