Switch over PEq Type instance to custom EqType family
Due to commit
http://git.haskell.org/ghc.git/commit/4be195e707d5e079dca5b6ce2e6cd34ff7624adf,
the next version of GHC will feature a new version of
Data.Type.Equality.== that leverages a closed type family definition
instead of the previous one (which was open). Alas, with this new
version of (==), it is no longer the case that
(forall a. (a == a) ~ True), as the new (==) cannot distinguish
between these two cases:

  f a == g b = f == g && a == b
  a   == a   = True

As a result, the SEq Type instance breaks on GHC HEAD, because it
assumes that (forall a. (a == a) ~ True). To avoid this breaking in
the next version of GHC, this commit switches the PEq Type instance
over to use a custom EqType type family, which encodes the previous
behavior of (==).

There's no change in behavior from this patch, but it will ensure
forwards compatibility.