Fix enormous oversight in singType
In my first iteration of this patch, I was under the mistaken
impression that all of the kind variables we'd need to explicitly
binds would be in the argument /types/. But this is completely
wrong—they can appear in the /return/ type, too! (Perhaps this is
obvious in hindsight, but it took me a while to dawn on this
realization.) The example that tipped me off to this can be found in
https://github.com/goldfirere/singletons/issues/296#issuecomment-363134457

While doing this, I discovered another spot where we weren't
explicitly binding kinds: the GADT type constructors used for Sing
instances. We now do so.