Fix #312 by not binding method tysig variables too early
Previously, we had an elaborate setup to store the variables bound by
a class (in `cd_bound_kvs`) and the variables bound by a class method
(in `lde_bound_kvs`) separately. But we were putting too many vars in
`lde_bound_kvs` in the first place, leading to #312! This makes the
distinction between `cd_bound_kvs` and `lde_bound_kvs` moot, as we
only want to bind the things that were put in `cd_bound_kvs`.

This presents a nice code simplification—we can axe `cd_bound_kvs`
entirely, and simply take what was previously put there and instead
put it in `lde_bound_kvs`. Much cleaner—and it actually works, too!

I'd like to add a more significant test case to demonstrate that
this is working properly, such as:

```haskell
class Foo a where
  bar :: forall b. a -> b -> b
  bar = h where
    h :: forall c. c -> b -> b
    h _ x = x
```

But it turns out that tickles
https://ghc.haskell.org/trac/ghc/ticket/14885, so this is off-limits
for the time being.