Fix #312 (again)
When singling default method implementations, it turns out we
generate a default type signature in which we apply a
defunctionalization symbol to /all/ of the type variables of the
singled class methods' type signature. But after commit
21777c5755e10d539a1ea884454c6654993d38e2, we introduced extra kind
variables in singled methods' type signatures, which we don't want
as arguments. Filtering them out does the trick, happily enough.