Use CmpNat, CmpSymbol to define the PEq instances for Nat, Symbol
Using Data.Type.Equality.== from base-4.11 means that (forall a. a == a ~ True)
will no longer hold for any type `a`. Unfortunately, the PEq instances for Nat
and Symbol happen to be implemented in terms of DTE.==, so this resulted
in an unfortunate regression in terms of the number of facts that can
be concluded about (a == a).

The same thing happened in
https://github.com/goldfirere/singletons/commit/d6d9f083fdfdc76529ef7708ad7f781a72a3af67,
but in that case, there was a test which broke. Unfortunately, no such tests
existed for `Nat` and `Symbol`, so this went unnoticed. I've added such a
test here.