Update supported constructors in README, and revise a validity check for constructors
Our README was a bit behind the times when it came to discussing
supported and unsupported features, so I brought it up to speed. In
particular, we (Richard and I) discovered recently that singletons
effectively does not support any constructors with contexts due to
the way SingKind is designed. Despite this, the code in singletons
was optimistically assuming that equality constraints were
permissible, but even that is untrue, so I removed a useless validity
check which detected the use of equality constraints.

Fixes #330.