Follow-up regression test for #312
In 21777c5755e10d539a1ea884454c6654993d38e2, I lamented the fact that
I couldn't add a regression test involving scoped type variables in
class methods due to Trac #14885. Luckily, that ticket has been
fixed, and since we now require GHC 8.6 as the minimum, it is safe
to assume that this fix is present.