Preemptively bump major version to 2.6
We technically haven't made any changes yet which would demand a
major version bump, but we will certainly do so by the time we
switch over to GHC 8.8. Until then, it's convenient to have a
separate version number for the development version of `th-desugar`.

This also bumps the `th-desugar` submodule to use version 1.10, and
requires that as the minimum in `singletons.cabal`.