Update th-desugar submodule
This commit looks large, but most of the changes are due to:

* Mechanically changing `Pred` to `Type`, now that the former is a synonym
  for the latter
* Mechanically changing `-Pa` to `-P`, now that `DPat` uses the latter suffix
  instead of the former

Along the way, I was able to delete a couple of extraneous functions (e.g.,
`substPred`). On the flip side, I had to add a couple of failure cases for
`DArrowT`/`DLitT` in functions that only accept `DPred`. Oh well.