Make singletons deterministic (#386)
There were several places in `singletons` that would compute entirely
different things depending on the order in which GHC's unique numbers
happened to be computed. Like the corresponding `th-desugar` patch
(in goldfirere/th-desugar#115), this fixes these issues by swapping
out nondeterministic uses of `Map` and `Set` with `th-desugar`'s new
`OMap` and `OSet` data structures, which remember the order in which
elements were inserted.

This patch looks large, but half of the modifications are routine
changes brought about by switching data structures, and the other
half are test suite wibbles brought about by `singletons` settling
on a deterministic order for the expected output. The upshot is that
we can finally run the `singletons` test suite with
`-dunique-increment=-1` and have it still pass! Hooray!

Fixes #367.