Make SomeSing's Enum efficient more efficient and general (#395)
It turns out that we don't need the `Enum (SomeSing k)` instance to
have a `SingKind k` constraint, as we can get away without needing
the use of `fromSing` in the implementations of `enumFromTo` and
`enumFromThenTo` if we convert an `SList @k` directly into a
list of `SomeSing k`s without recursively invoking `fromSing`. As an
added bonus, this will likely make the code more efficient at runtime
as well.