Minor simplification in D.S.P.Defun.defunctionalize
According to a careful reading of the code in
`Data.Singletons.Promote.Defun.defunctionalize`, the kind of
`tyfun_param` should, allegedly, be of the form `a ~> b` (`m_tyfun`),
where `tyfun_param` is the last type parameter in defunctionalization
data types of the following form:

```hs
data FSymN (a1 :: k1) ... (an :: kn) tyfun_param where ...
```

When I first saw this, I wondered how this could possibly
kind-check, since if `FSymN :: k1 -> ... kn -> a ~> b`, then the kind
of `tyfun_param` should be `TyFun a b`, not `a ~> b` (since
`a ~> b` expands to `TyFun a b -> Type`). But upon staring at the
code further, I realized that the kind of `tyfun_param` is never
used! That is because if we're in a situation where the return kind
of `FSymN` is known, `tyfun_param` will never be used. Instead, this
will be generated:

```hs
data FSymN :: forall k1 ... kn a b. k1 -> ... kn -> a ~> b where ...
```

Indeed, the only situation where `tyfun_param` is actually used is
when we /don't/ know the return kind of `FSymN`, in which
`singletons` generates a declaration for `FSymN` /without/ an
explicit kind for `tyfun_param`. Therefore, this bug went unnoticed
by sheer accident.

Since any attempt to give `tyfun_param` is wasted effort anyway,
let's just not do so.