Make qNewUnique return a Uniq instead of an Int
After GHC commit
https://gitlab.haskell.org/ghc/ghc/commit/4ba73e00c4887b58d85131601a15d00608acaa60,
`Uniq` is an `Integer` instead of `Int`. The `qNewUnique` function,
however, currently returns an `Int` by using `fromIntegral` to
convert the `Integer` to an `Int`. This is potentially dangerous,
however, as this could truncate large `Integer`s down into smaller
`Int`s, which has the potential to return non-unique `Int`s.

Instead of doing this, let's simply make `qNewUnique` return a `Uniq`
without any modification. The only place where `singletons` uses
`qNewUnique` is for the purpose of turning the `Uniq` directly into
a `String` anyway, so this is a perfectly fine thing to do.

This is originally Matthew Bauer's patch from #398, which was an
abandoned attempt at fixing a separate GHCJS bug. Since this patch is
useful on its own merits, I (@RyanGlScott) have modified the commit
message to reflect its new purpose.

Co-authored-by: Matthew Bauer <mjbauer95@gmail.com>
Co-authored-by: Ryan Scott <ryan.gl.scott@gmail.com>