Don't pick up .ghc.environment files when running tests (#399)
As observed in commercialhaskell/stackage#4646, newer versions of
`stack` use `.ghc.environment` files in such a way that they are
picked up while invoking the test suite, which causes a
`Loaded package environment from ...` message to be printed and
breaks the expected test output. To avoid this issue, we simply
pass `-package-env=-` when shelling out to `ghc` in the test suite
so that it never picks up any ambient `.ghc.environment` files.