Fix #404
As noted in #404, there were situations where the `-ddump-splices`
output of `singletons` would print out longer suffixes for some
occurrences of a name that would not appear for other occurrences of
the same name. The culprit is that the latter occurrences of the name
were the output of `promoteValNameLhs` while the former were not.

The solution adopted here is to use the latter style everywhere.
That is, we ensure that the name is always fed into
`promoteValNameLhs` before being printed out. It turns out there were
two places that were not doing this: `promoteLetDecRHS` and
`defunctionalize`.

There is no test case only because the difference only manifests when
`-ddump-splices` prints out uniques. Since we run the testsuite with
`-dsuppress-uniques` for determinism reasons, this would not show up
when running tests.