Properly reject higher-rank types during promotion/singling (#406)
* Properly reject higher-rank types during promotion/singling

`singletons` can't handle higher-rank types. You might think that
these sorts of types would naturally be rejected when promoting or
singling them, but that actually turns out not to be the case!
(See #401 for examples of higher-rank types that squeeze through the
cracks.)

Luckily, the additional machinery that `th-desuguar` adds in
goldfirere/th-desugar#125 makes it much easier to detect these.
Instead of using `unravelDType` to decompose a function type (which
permits all sorts of higher-rank gubbins), we use a restricted
version called `unravelVanillaDType`, which throws an error if
anything higher-rank is detected. See
`Note [Vanilla-type validity checking during promotion]` in
`D.S.Promote.Type` for a full explanation of what a "vanilla" type
is.

Fixes #401.

* Review comments

* Add unravelVanillaDType_either

* Use correct th-desugar commit