Fix #410 (#411)
The only part of this patch that actually changes the behavior of
`singletons` is the new `Map Name [DTyVarBndr]` argument of
`promoteInstanceDec`, which maps names of type classes to the type
variable binders in their class heads. This map is consulted first
when constructing a substitution from class variables to instance
types so as to avoid imported classes from shadowing them, as
observed in #410.

The rest of this patch simply refactors some of the structure of the
related functions `promoteMethod` and `promoteLetDecRHS`, both of
which have arguments that are only used in certain code paths
depending on what sort of thing is being promoted. To make this
clearer, I introduced auxiliary data types `MethodSort` and
`LetDecRHSSort` that describe the various code paths.

Fixes #410.