Make the test suite less awk-ward (#419)
This simplifies the design of the test suite by:

1. Removing the use of `awk`, as it is not needed, and
2. Remove the use of `.ghcXX.template` files in favor of ordinary
   `.golden` files.

This allows us to replace the current tailor-made (and buggy)
`tasty-golden` setup with a more conventional one and delete lots
of code in the process. I've also made an effort to clean up the
documentation in `tests/README.md` and the Notes in
`SingletonsTestSuiteUtils` to better reflect the current design.

Fixes #417.