Refactor promoteLetDecRHS to be less crufty
`promoteLetDecRHS` has two cases: one for `UValue` and `UFunction`.
Both of these cases duplicate a significant amount of code, which I
find hard to follow. This patch reduces the code duplication as much
as possible to reveal the underlying structure.

Along the way, I noticed that `promoteClause` and `promoteMatch`
were computing `allLocals` wastefully, as this information is
available at every call site of these functions. I tweaked the
types of these functions to accept a full `DType` consisting of
the promoted `Name` applied to its local arguments (instead of
accepting only a `Name` and then recomputing its local arguments
with `allLocals`).

This patch is purely refactoring, so there should be no change in
behavior.