Remove special case for [] in promoteValRhs
The comment claims that this case was needed at one point to work
around #21, but that no longer appears to be the case. Some test
cases' expected output changes due to occurrences of `'[]` changing
to `NilSym0`. To make the `Singletons.Nat` test case work in
particular, we now reexport `NilSym0` from `Data.Singletons.TH`,
just as the defunctionalization symbols for `(:)` are reexported.

Fixes #428.