Simplify DLamE case of promoteExp slightly
The `DLamE` case of `promoteExp` was generating two sets of argument
names: one for the promoted type family's type variable
binder names, and another for the argument names in the type family
equation. Since the type family only has one equation, however, we
can just generate one set of argument names and use it in both
places.

Predictably, some tests' expected outputs alter slightly as a result
of this change.