Introduce Options and OptionsMonad (#427)
* Bump major version to 2.7

* Introduce Options and OptionsMonad

This patch introduces an `Options` data type and an `mtl`-like
`OptionsMonad` class for monads that carry `Options`. At present,
the only things one can do with `Options` are:

* Toggle the generation of `SingKind` instances. Suppressing
  `SingKind` instances provides an effective workaround for #150.
* Hook into the TH machinery's naming conventions for promoted and
  singled names. This fixes #204.

The vast majority of this patch simply adds plumbing by using
`OptionsMonad` in places that need it. See the `D.S.TH.Options`
module for where most of the new code is housed, as well as the
`T150` and `T204` test cases for examples of how to use it.