Derive some instances for QWithAux, SgM using GeneralizedNewtypeDeriving
The Quasi instances for QWithAux and SgM are very lengthy and require
lots of code churn to keep updated whenever new methods are added to
Quasi. This is a shame, as these instances don't really do anything
interesting—they just inherit the behavior of the underlying monad
transformers! It took me a while to realize this since the `qRecover`
implementations _appear_ to be custom-made, but if you squint, you'll
see that they just inline the definition of `qRecover` for `WriterT`.

In light of this, let's save ourselves from future code churn by
deriving these instances with `GeneralizedNewtypeDeriving`. We can
even do the same for `QWithAux`'s `DsMonad` instance, so let's throw
that in for an added bonus.