Add Data.Singletons.Prelude.Proxy
Even in a library as deeply entrenched in type-level trickery as
`singletons` is, there are still some situations that call for
proxies. See, for example, the use case in #435, which uses not only
the classic `Proxy` data type, but also the /singled/ version of
`Proxy`! The author of this use case requested that the singled
version of `Proxy` be made available in `singletons`, so this patch
accomplishes that.

Fixes #435.