Adapt to simplified subsumption
GHC's
[simplified subsumption proposal](https://github.com/ghc-proposals/ghc-proposals/blob/master/proposals/0287-simplify-subsumption.rst)
removes deep skolemization, which `singletons`' TH machinery relied
on in order to typecheck the uses of `singFun1` _et al._ that it
generates. The reason it relied on deep skolemization is because
`SingFunction1` _et al._ were defined with `forall`s nested
underneath arrow types. In this patch, we redefine `SingFunction1`
_et al._ so that all of the `forall`s are up front, making deep
skolemization unnecessary when applying `singFun1` _et al._

...well, unnecessary 99% of the time. It turns out there is one
place in `ByHand` that defines a singled function with nested
`forall`s that no longer works with the new version of `singFun2`.
But `singletons` would never generate a function with nested
`forall`s like this in the first place, so let's just change it to
use prenex `forall`s.

Fixes #438.