Disable kind generalization when singling sig-less local functions (#440)
When a local function (1) lacks a type signature, and (2) has no
arguments, we single it such that its type signature has a return
type of `Sing @_ Blah` instead of `Sing Blah`. This prevents kind
generalization from mucking with the kind of `Blah`, which can be
undesirable. For the complete story, refer to
`Note [Disable kind generalization for local functions if possible]`
in `D.S.Single`.

Aside from fixing #296, this allows us to finally uncomment a line
of code in `T183` that relies on this fix.