Widen lower version bounds on Cabal in custom-setup
`singletons` crucially relies on the latest versions of `base` and
`template-haskell`, but there is no reason to require the absolute
latest version of `Cabal` in `singletons`' custom setup script, as
it is backwards-compatible with `Cabal-3.0`. Moreover, there is a
good reason to permit building with `Cabal-3.0`: building
`singletons` with `cabal-install-3.0` requires it. Since
`cabal-install-3.2` has not been released yet, there are some
environments (e.g., `head.hackage`) where `cabal-install-3.0` is
still the only option, so this patch supports building `singletons`
in those environments.