Declare fully saturated defun symbols as type families
The most immediate benefit of doing so is that it fixes #445, a UX
regression involving GHCi's `:kind!` command. It also has the benefit
of allowing fully saturated defunctionalization symbols to have
standalone kind signatures, which makes much of
`Note [No SAKs for fully saturated defunctionalization symbols]`
moot. As a result, I have replaced that Note with the new
`Note [Fully saturated defunctionalization symbols]`.