More thoroughly document how to work with datatypes that use arrows/literals (#451)
Previously, the only mention of how to do so in the `README` was a simple link
to #76. But finding the relevant information in #76 requires digging through
several comments, and to make things worse, the actual advice (in
https://github.com/goldfirere/singletons/issues/76#issuecomment-44220892)
has bitrotted!

This patch (1) cleans up the bitrot, and (2) enshrines the advice in
the `README` itself, where it belongs.

Fixes #76 (again).