Fix #453 by always explicitly quantifying SAK kind variables
If a data declaration's SAK were to use an explicit `forall`, then
`singletons` would have no issues singling it. If it didn't use an
explicit `forall`, however, then it would topple over. The fix:
quantify the kind variables in a SAK ourselves when a SAK lacks an
explicit `forall`. The `D.S.Single.Type.singTypeKVBs` function does
the exact same thing, as it turns out, but I somehow overlooked this
before.

While I was in town, I took the liberty of cleaning up a couple of
code smells:

* There were a couple of invocations of `promoteType` that were
  completely unnecessary.
* In the case where a data declaration lacks a SAK, I now explicitly
  quantify over kind variables as well as type variables. Not
  quantifying over kind variables in a data declaration's return kind
  is not an error per se (since the forall-or-nothing rule doesn't
  apply there), but it is more consistent with what happens with
  data declarations that do have SAKs.