Use simpler version of getRecordSelectors from th-desugar (#456)
This patch updates the `th-desugar` submodule to incorporate
goldfirere/th-desugar#139, which:

1. Removes the horribly buggy `conExistentialTvbs` function, and
2. Simplifies `getRecordSelectors` so that it does not bother trying
   to filter out naughty record selectors.

Incorporating (2) amounts to fixing #443, for which the test case was
rejected due to `conExistentialTvbs` being overly conservative with
GADTs. To adapt to (1), certain validity checks in `singletons` which
previously used `conExistentialTvbs` will now be less precise. This
isn't a huge deal—it just means that erroneous programs which
`singletons` used to give specialized error messages for will now
just go through, resulting in an error message on GHC's behalf.

Fixes #443.