Make {P,S}Eq less magical, don't make DefaultEq the default (#458)
Currently, the `PEq` and `SEq` classes are implemented by hand in
`D.S.Prelude.Eq` due to the default implementation, `DefaultEq`. But
as discussed in #457, this default isn't actually wise for most
singleton types. This patch removes `DefaultEq` as the default, which
allows `PEq` and `SEq` to be generated with Template Haskell, just
like other classes in the `D.S.Prelude.*` module namespace. As a
result, this fixes #457.

This refactoring led me to discover some other unusual treatment of
`PEq` and `SEq`. Namely, derived instances of `PEq` and `SEq` are
very magical in that they are handled entirely outside of
`D.S.Partition.partitionDecs`. But there isn't much of a good reason
to do so—we can just as well generate a normal `Eq` instance, then
promote and single it using the existing TH machinery. The only
magical thing about deriving `Eq` is that it also produces
`SDecide`/`TestEquality`/`TestCoercion` instances, but that can be
done independently of `PEq` and `SEq`.

This patch also refactors the treatment of `deriving Eq` to remove
the special casing for `PEq` and `SEq`:

* The `D.S.Promote.Eq` and `D.S.Single.Eq` modules are now gone,
  having been subsumed by the new `D.S.Deriving.Eq` module, which
  generates an instance just like `singletons` would for any other
  `stock`-derivable class. What remains of `D.S.Single.Eq` is now
  located in `D.S.Single.Decide`, which only exists to generate
  instances of `SDecide`, `TestEquality`, and `TestCoercion`.
* `D.S.TH` had strange functions `singEqInstanceOnly` and
  `singEqInstancesOnly` that generated instances of `SEq`, but not
  `PEq`. No other `stock`-derivable class had `-Only` variants like
  this, and given that `singEqInstance{s}` already exist, they were
  practically useless. I decided to take the opportunity to remove
  these warts.