Generate redundant kind information for Haddock's benefit
Unfortunately, a Haddock limitation (discussed in #466) causes
standalone kind signatures not to be rendered at all, which makes the
Haddocks for much of `singletons-base` omit crucial kind information.
To work around this Haddock limitation, this patch generates extra
argument and result kinds in the bodies of type-level declarations.
See `Note [Keep redundant kind information for Haddocks]` in
`D.S.Promote`.