Rename modules in singletons-base
This renames various modules in `singletons-base` to use a naming
convention that better reflects the parts of `base` that each module
is intended to reflect. See the `singletons-base` `CHANGES.md` file
for a full breakdown of what was changed.

This is the second part of a two-commit series that fixes #464.