Make Prelude.Singletons mirror the Prelude more closely
A recent perusal of `Prelude.Singletons` revealed that there were
several functions from the `Prelude` that `Prelude.Singletons` didn't
export, and conversely, there were various functions that
`Prelude.Singletons` does export that the `Prelude` doesn't! Part of
the reason why this went unnoticed for so long is that the structure
of `Prelude.Singletons`' exports is rather haphazard, so attempting
to eyeball the differences between the export lists of
`Prelude.Singletons` and the `Prelude` was unnecessarily difficult.

This cleans things up by restructuring `Prelude.Singletons`' export
list to very closely mirror the `Prelude`'s export list. For each
exported item in the `Prelude` that `Prelude.Singletons` _doesn't_
export, a comment has been left in the corresponding place in
`Prelude.Singletons`. This made it easy to tell which things should
(and shouldn't) have been exported from the `Prelude.Singletons`. All
such changes have been recorded in the `CHANGES.md` file.

Fixes #465.