Enable FlexibleContexts in more places to support GHC 9.2
`UndecidableInstances` no longer implies `FlexibleContexts` in instance
declarations as of GHC 9.2 (see
https://gitlab.haskell.org/ghc/ghc/-/wikis/migration/9.2#undecidableinstances-no-longer-implies-flexiblecontexts-in-instance-declarations),
so we need to enable `FlexibleContexts` in a handful of additional places to
make `singletons` compile with GHC 9.2. Luckily, this is a backwards-compatible
change, so I decided to just make this change now.

(Originally noticed in `head.hackage` in
https://gitlab.haskell.org/ghc/head.hackage/-/merge_requests/138.)