Tweak ShowSing' so that Show instances for singletons can be derived (#486)
By giving `ShowSing'` a quantified superclass constraint, it becomes possible
to derive `Show` instances for singleton types that mention `ShowSing` in the
instance context. This is a remarkably backwards-compatible change except for
one corner case, which I have documented in the Haddocks for `ShowSing'` in
`Data.Singletons.ShowSing`.

Fixes #426, as well as one bullet point of #439.