Bump th-desugar to incorporate VTA in data constructor patterns
This bumps the `th-desugar` submodule to commit
320453f4b21efb8040226c4fe54ebb088e902657, which adds support for visible type
applications in data constructor patterns. Note that this does not add support
for promoting or singling VTA in patterns yet—we leave that to future work.
For now, `singletons-th` will emit a warning if it encounters a data
constructor pattern with one or more visible type arguments during promotion,
much like it currently does when it encounters a visible type argument in an
expression.