Monomorphize (%<=?) and friends to Natural
GHC 9.2 generalizes the kind of `(<=?)` from `Natural` to being polymorphic,
which confuses `singletons-base`. Fixing the issue is as simple as adding some
explicit `Natural` kinds to the corresponding definitions in `singletons-base`,
such as `(%<=?)`.

We may wish to consider implementing a more general version of `(%<=?)` in the
future to reflect its more general kind in `base`. For now, however, this will
suffice.