singletons-th: Support type applications in data constructor patterns
While supporting visible type applications in their full generality is
challenging, type applications in data constructor patterns are a very well
behaved subset, which makes it quite feasible to support in `singletons-th`.
This patch accomplishes just that.

We also make an effort to distinguish between wildcard types found in type
applications in data constructor, which we can support, and other wildcard
types, which we cannot yet support due to GHC restrictions about using wildcard
types in kind-level contexts. I have added a section to the `README` to explain
this distinction.

Fixes #489.