Fix #536 with tactical uses of promoteType
The code paths for generating `SDecide` and `ShowSing` instances are somewhat
different from the code paths for other derived instances, and while the latter
code paths remembered to call `promoteType` on types that are meant to appear
as kinds, the `SDecide` and `ShowSing` code paths did not remember to do this.
Easily fixed.