singletons-th: Remove unused `unSingFun*` machinery
Previous versions of `singletons` generated uses of `unSingFun*` as a way to
accomplish eta expansion. In the current version of the library, however, eta
expansion is instead accomplished by giving function clauses more patterns,
which removes the need for `unSingFun*` entirely. There were still some
vestigial remains of the old `unSingFun*` machinery lying around, however,
so this patch removes them.

While I was in the neighborhood, I took the opportunity to document some of the
invariants that the current eta expansion machinery relies on.

Fixes #545.