Greatly simplify how implicitly quantified type signatures are singled
This patch rips out essentially all of the implicit-type-variable-tracking
machinery described in the former `Note [Explicitly binding kind variables]` in
favor of a much simpler approach that relies on GHC to perform implicit
quantification. See the newly revamped `Note [Preserve the order of type
variables during singling]` (Wrinkle 1: Dealing with the presence (and absence)
of `forall`) in `D.S.TH.Single.Type` for the full details.

Fixes #550.