Use `(<>)` in derived Ord instances, remove `thenCmp`
Previously, the code that `singletons-th` would generate when promoting or
singling a derived `Ord` instance would use a bespoke `thenCmp :: Ordering ->
Ordering -> Ordering` function.  This caused two issues:

1. `Data.Ord.Singletons` had to re-export `thenCmp`, despite the fact that no
   such function exists in `base`.
2. Users would have to remember to import this function to support derived
  `Ord` instances with `singletons-th`.  Failing to remember to do this would
  result in confusing errors like the one encountered in #555.

To fix both issues, this patch removes `thenCmp` in favor of `(<>) @Ordering`.
Happily, the promoted and singled `Semigroup` classes are exported by
`Prelude.Singletons`, so this means that derived `Ord` instances will now have
everything that they need just from importing `Data.Singletons.TH` and
`Prelude.Singletons`. (It is unclear if this is something that we can guarantee
to work for all possible derived instances, but it is nice to do when we can.)

Fixes #555.